+++ hummingbot/core/data_type/OrderExpirationEntry.pxd
@@ -2,7 +2,7 @@
 
 from libcpp.string cimport string
 
-cdef extern from "cpp/OrderExpirationEntry.h":
+cdef extern from "../cpp/OrderExpirationEntry.h":
     cdef cppclass OrderExpirationEntry:
         OrderExpirationEntry()
         OrderExpirationEntry(string symbol,
