+++ hummingbot/core/data_type/order_book_tracker.py
@@ -22,7 +22,7 @@ from .order_book_message import (
     OrderBookMessageType,
     OrderBookMessage,
     )
-from hummingbot.market.data_source.order_book_tracker_data_source import OrderBookTrackerDataSource
+from hummingbot.core.data_type.order_book_tracker_data_source import OrderBookTrackerDataSource
 
 TRADING_PAIR_FILTER = re.compile(r"(BTC|ETH|USDT)$")
