+++ hummingbot/core/data_type/remote_api_order_book_data_source.py
@@ -18,7 +18,7 @@ from websockets.exceptions import ConnectionClosed
 
 import conf
 from hummingbot.market.binance.binance_order_book import BinanceOrderBook
-from hummingbot.market.data_source.order_book_tracker_data_source import OrderBookTrackerDataSource
+from hummingbot.core.data_type.order_book_tracker_data_source import OrderBookTrackerDataSource
 from hummingbot.logger import HummingbotLogger
 from hummingbot.core.data_type.order_book_tracker_entry import OrderBookTrackerEntry
