+++ hummingbot/core/data_type/user_stream_tracker.py
@@ -7,7 +7,7 @@ import logging
 from typing import (
     Optional
 )
-from hummingbot.market.data_source.user_stream_tracker_data_source import UserStreamTrackerDataSource
+from hummingbot.core.data_type.user_stream_tracker_data_source import UserStreamTrackerDataSource
 from hummingbot.logger import HummingbotLogger
