+++ hummingbot/market/bamboo_relay/bamboo_relay_api_order_book_data_source.py
@@ -18,7 +18,7 @@ from websockets.exceptions import ConnectionClosed
 
 from hummingbot.market.bamboo_relay.bamboo_relay_order_book import BambooRelayOrderBook
 from hummingbot.market.bamboo_relay.bamboo_relay_active_order_tracker import BambooRelayActiveOrderTracker
-from hummingbot.market.data_source.order_book_tracker_data_source import OrderBookTrackerDataSource
+from hummingbot.core.data_type.order_book_tracker_data_source import OrderBookTrackerDataSource
 from hummingbot.core.data_type.order_book_tracker_entry import OrderBookTrackerEntry, BambooRelayOrderBookTrackerEntry
 from hummingbot.core.data_type.order_book_message import OrderBookMessage, BambooRelayOrderBookMessage
 from hummingbot.core.utils.exchange_rate_conversion import ExchangeRateConversion
