+++ hummingbot/market/binance/binance_api_order_book_data_source.py
@@ -19,7 +19,7 @@ from websockets.exceptions import ConnectionClosed
 from hummingbot.market.binance.binance_order_book import BinanceOrderBook
 from hummingbot.core.utils import async_ttl_cache
 from hummingbot.logger import HummingbotLogger
-from hummingbot.market.data_source.order_book_tracker_data_source import OrderBookTrackerDataSource
+from hummingbot.core.data_type.order_book_tracker_data_source import OrderBookTrackerDataSource
 from hummingbot.core.data_type.order_book_tracker_entry import OrderBookTrackerEntry
 from hummingbot.core.data_type.order_book_message import OrderBookMessage
