+++ hummingbot/market/coinbase_pro/coinbase_pro_api_order_book_data_source.py
@@ -17,7 +17,7 @@ import websockets
 from websockets.exceptions import ConnectionClosed
 
 from hummingbot.market.coinbase_pro.coinbase_pro_order_book import CoinbaseProOrderBook
-from hummingbot.market.data_source.order_book_tracker_data_source import OrderBookTrackerDataSource
+from hummingbot.core.data_type.order_book_tracker_data_source import OrderBookTrackerDataSource
 from hummingbot.core.utils import async_ttl_cache
 from hummingbot.logger import HummingbotLogger
 from hummingbot.core.data_type.order_book_tracker_entry import (
