+++ hummingbot/market/coinbase_pro/coinbase_pro_api_user_stream_data_source.py
@@ -11,7 +11,7 @@ from typing import (
 import ujson
 import websockets
 from websockets.exceptions import ConnectionClosed
-from hummingbot.market.data_source.user_stream_tracker_data_source import UserStreamTrackerDataSource
+from hummingbot.core.data_type.user_stream_tracker_data_source import UserStreamTrackerDataSource
 from hummingbot.market.coinbase_pro.coinbase_pro_auth import CoinbaseProAuth
 from hummingbot.logger import HummingbotLogger
 from hummingbot.core.data_type.order_book_message import OrderBookMessage
