+++ hummingbot/market/coinbase_pro/coinbase_pro_user_stream_tracker.py
@@ -6,7 +6,7 @@ from typing import (
     Optional,
     List,
 )
-from hummingbot.market.data_source.user_stream_tracker_data_source import UserStreamTrackerDataSource
+from hummingbot.core.data_type.user_stream_tracker_data_source import UserStreamTrackerDataSource
 from hummingbot.logger import HummingbotLogger
 from hummingbot.core.data_type.user_stream_tracker import (
     UserStreamTrackerDataSourceType,
