+++ hummingbot/market/ddex/ddex_api_order_book_data_source.py
@@ -19,7 +19,7 @@ from websockets.exceptions import ConnectionClosed
 from hummingbot.core.utils import async_ttl_cache
 from hummingbot.market.ddex.ddex_active_order_tracker import DDEXActiveOrderTracker
 from hummingbot.market.ddex.ddex_order_book import DDEXOrderBook
-from hummingbot.market.data_source.order_book_tracker_data_source import OrderBookTrackerDataSource
+from hummingbot.core.data_type.order_book_tracker_data_source import OrderBookTrackerDataSource
 from hummingbot.logger import HummingbotLogger
 from hummingbot.core.data_type.order_book_tracker_entry import (
     DDEXOrderBookTrackerEntry,
