+++ hummingbot/market/ddex/ddex_order_book_tracker.py
@@ -19,8 +19,8 @@ from hummingbot.core.data_type.order_book_tracker import (
     OrderBookTrackerDataSourceType
 )
 from hummingbot.market.ddex.ddex_order_book import DDEXOrderBook
-from hummingbot.market.data_source.order_book_tracker_data_source import OrderBookTrackerDataSource
-from hummingbot.market.data_source.remote_api_order_book_data_source import RemoteAPIOrderBookDataSource
+from hummingbot.core.data_type.order_book_tracker_data_source import OrderBookTrackerDataSource
+from hummingbot.core.data_type.remote_api_order_book_data_source import RemoteAPIOrderBookDataSource
 from hummingbot.market.ddex.ddex_api_order_book_data_source import DDEXAPIOrderBookDataSource
 
 from hummingbot.core.data_type.order_book_message import (
